v0.08
- Added a mouse-click effect on ground
- Changed terrain to Convex to support instantiation at ray hit point for mouse effect
- Fixed a frame-drop issue caused by spamming abilities after changing terrain to Convex by removing GetCanvasRotation debug message
- Added an AlertObserver to projectiles.
    - Projectile hitboxes can now be adjusted through passing in Collider Start and End in the Animation Timeline
- Added a warning target area hitbox to the W ability (this will be instantiated over the network to show enemy players where the ability was targeted)
- Fixed an issue where the Q projectile did not reach its desired location by moving its RigidBody physics movement script into FixedUpdate
- Scaled down In-Game Health and Mana Bars over Player Character

- Added Mana Costs to Q and W abilities
- Fixed Mana Bar visual bug
- Added System Notification Popup message with AlertObserver
- Refactored AlertObservers to Switch Statements

- Added Mana Regen and clamped whole numbers
- Added Health Regen and clamped whole numbers
- Fixed a bug where the Q single-target projectile hits multiple enemies if tightly grouped together by adding a Hit Bool.
    - Similar application for W ability. Hit Bool should not be called for AOE abilities likewise.

+++ Assets/Scripts/Character/Stats Init/AbilityClass.cs
@@ -18,7 +18,7 @@ public class AbilityClass
     public string abilityName;
 
     public int abilityLevel;
-    public int abilityCost;
+    public float abilityCost;
 
     public float abilityCooldown;
     public float abilityDuration;
